small refactor of services.py
