+++ app/core/config.py
@@ -12,4 +12,4 @@ class Config(BaseSettings):
     model_config = SettingsConfigDict(env_file=".env", extra="allow")
 
 
-settings = Config()
+config = Config()
